tests/sensors/Accelerometer.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Accelerometer Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://w3c.github.io/accelerometer/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./resources/generic-sensor-tests.js"></script>
<div id="log"></div>
<script>

function verify_sensor_reading(sensor, should_be_bull) {
  if (should_be_bull) {
    return sensor.x == null &&
           sensor.y == null &&
           sensor.z == null &&
           sensor.timestamp == null;
  }
  return sensor.x != null &&
         sensor.y != null &&
         sensor.z != null &&
         sensor.timestamp != null;
}

function verify_both_sensor_reading(sensor1, sensor2) {
  return sensor1.x == sensor2.x ||
         sensor1.y == sensor2.y ||
         sensor1.z == sensor2.z ||
         sensor1.timestamp == sensor2.timestamp;
}

async_test(t => {
  let sensor1 = new Accelerometer({includeGravity: false});
  let sensor2 = new Accelerometer();
  sensor1.onactivate = t.step_func_done(() => {
    assert_false(verify_both_sensor_reading(sensor1, sensor2));
    sensor1.stop();
    sensor2.stop();
  });
  sensor1.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  sensor2.onerror = t.step_func_done(event => {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
  sensor1.start();
  sensor2.start();
}, "when construct Accelerometer with includeGravity that is set to false, the reading is different from the default one");

runSensorFrequency(Accelerometer);
runGenericSensorTests(Accelerometer, verify_sensor_reading);

</script>
